Gave Custom Tag responsibility to handle how to convert from `PBIntermediateNode` to `PBEgg`
