+++ lib/eggs/custom_egg.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 
+import 'package:directed_graph/directed_graph.dart';
 import 'package:parabeac_core/controllers/main_info.dart';
 import 'package:parabeac_core/generation/generators/import_generator.dart';
 import 'package:parabeac_core/generation/generators/pb_generator.dart';
@@ -7,6 +8,8 @@ import 'package:parabeac_core/generation/generators/plugins/pb_plugin_node.dart'
 import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/commands/write_symbol_command.dart';
 import 'package:parabeac_core/generation/generators/value_objects/file_structure_strategy/file_ownership_policy.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/interfaces/pb_injected_intermediate.dart';
+import 'package:parabeac_core/interpret_and_optimize/entities/pb_shared_instance.dart';
+import 'package:parabeac_core/interpret_and_optimize/entities/pb_shared_master_node.dart';
 import 'package:parabeac_core/interpret_and_optimize/helpers/child_strategy.dart';
 import 'package:parabeac_core/interpret_and_optimize/helpers/pb_context.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/subclasses/pb_intermediate_node.dart';
@@ -36,6 +39,64 @@ class CustomEgg extends PBEgg implements PBInjectedIntermediate {
       frame,
       originalRef.name.replaceAll('<custom>', '').pascalCase,
     );
+  }
+
+  /// Handles `iNode` to convert into a [CustomEgg].
+  ///
+  /// Returns the [PBIntermediateNode] that should go into the [PBIntermediateTree]
+  PBIntermediateNode handleIntermediateNode(
+    PBIntermediateNode iNode,
+    PBIntermediateNode parent,
+    CustomEgg tag,
+    PBIntermediateTree tree,
+  ) {
+    if (iNode is PBSharedMasterNode) {
+      iNode.name = iNode.name.replaceAll('<custom>', '');
+      var tempGroup = CustomEgg(
+        null,
+        iNode.frame,
+        iNode.name.pascalCase + 'Custom',
+      );
+
+      tree.addEdges(
+          tempGroup, tree.childrenOf(iNode).cast<Vertex<PBIntermediateNode>>());
+
+      tree.replaceChildrenOf(iNode, [tempGroup]);
+      return iNode;
+    } else if (iNode is PBSharedInstanceIntermediateNode) {
+      var tempGroup = CustomEgg(
+        null,
+        iNode.frame,
+        tag.name + 'Custom',
+      );
+
+      iNode.parent = parent;
+
+      tree.replaceNode(iNode, tempGroup);
+
+      tree.addEdges(tempGroup, [iNode]);
+
+      return tempGroup;
+    } else {
+      // [iNode] needs a parent and has not been added to the [tree] by [tree.addEdges]
+      iNode.parent = parent;
+      // If `iNode` has no children, it likely means we want to wrap `iNode` in [CustomEgg]
+      if (tree.childrenOf(iNode).isEmpty) {
+        // Generate new [CustomEgg] with a new UUID to prevent cycles.
+        //? TODO: should every [CustomEgg] have unique UUID or inherit from iNode like it is currently doing in `generatePluginNode`?
+        var newTag = CustomEgg(
+          null,
+          iNode.frame,
+          tag.name,
+        );
+        tree.removeEdges(iNode.parent, [iNode]);
+        tree.addEdges(newTag, [iNode]);
+        return newTag;
+      }
+      tree.replaceNode(iNode, tag, acceptChildren: true);
+
+      return tag;
+    }
   }
 }
